change handling multiple same ledgers in LedgerConsensus
Necessary due to how LedgerExchange operates: the winner is returned. So a server might end up with two same ledgers, this implies that some other server already decided for us.

Possible improvement: use of MAC's for the winners. How will the architecture need to be adapted for accomodate for this?